Move helps to .utils/utils.js
